+++ workspaces/bulk-import/plugins/bulk-import/src/components/AddRepositories/AddRepositoriesForm.test.tsx
@@ -12,20 +12,26 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */ import React from 'react';
+ */
+ import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 import { configApiRef, identityApiRef } from '@backstage/core-plugin-api';
 import { MockConfigApi, TestApiProvider } from '@backstage/test-utils';
 
 import { useDrawer } from '@janus-idp/shared-react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { render, screen } from '@testing-library/react';
 import { useFormikContext } from 'formik';
 
 import { bulkImportApiRef } from '../../api/BulkImportBackendClient';
 import { mockGetImportJobs, mockGetRepositories } from '../../mocks/mockData';
-import { ImportJobStatus, RepositorySelection } from '../../types';
-import { AddRepositoriesForm } from './AddRepositoriesForm';
+import {
+  ImportJobStatus,
+  OrgAndRepoResponse,
+  RepositorySelection,
+} from '../../types';
+import { AddRepositories } from './AddRepositories';
 
 jest.mock('formik', () => ({
   ...jest.requireActual('formik'),
@@ -57,6 +63,16 @@ jest.mock('@mui/material', () => ({
   },
 }));
 
+const createTestQueryClient = () =>
+  new QueryClient({
+    defaultOptions: {
+      queries: {
+        retry: false, // Disable retries for testing
+      },
+    },
+  });
+let queryClient: QueryClient;
+
 class MockBulkImportApi {
   async getImportAction(
     repo: string,
@@ -65,6 +81,13 @@ class MockBulkImportApi {
     return mockGetImportJobs.imports.find(
       i => i.repository.url === repo,
     ) as ImportJobStatus;
+  }
+  async dataFetcher(
+    _pageNo: number,
+    _size: number,
+    _searchString: string,
+  ): Promise<OrgAndRepoResponse> {
+    return mockGetRepositories;
   }
 }
 
@@ -83,9 +106,10 @@ beforeEach(() => {
     },
     setFieldValue: jest.fn(),
   });
+  queryClient = createTestQueryClient();
 });
 
-describe('AddRepsositoriesForm', () => {
+describe('AddRepositoriesForm', () => {
   it('should render the repositories list with the footer', async () => {
     (useDrawer as jest.Mock).mockImplementation(initial => ({
       initial,
@@ -110,7 +134,9 @@ describe('AddRepsositoriesForm', () => {
             ],
           ]}
         >
-          <AddRepositoriesForm error={null} />
+          <QueryClientProvider client={queryClient}>
+            <AddRepositories error={null} />
+          </QueryClientProvider>
         </TestApiProvider>
       </Router>,
     );
@@ -147,9 +173,9 @@ describe('AddRepsositoriesForm', () => {
             ],
           ]}
         >
-          <AddRepositoriesForm
-            error={{ message: 'error', title: 'error occurred' }}
-          />
+          <QueryClientProvider client={queryClient}>
+            <AddRepositories error={{ err: 'error occurred' }} />
+          </QueryClientProvider>
         </TestApiProvider>
       </Router>,
     );