+++ workspaces/bulk-import/plugins/bulk-import/src/components/Repositories/CatalogInfoAction.tsx
@@ -12,7 +12,8 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */ import React from 'react';
+ */
+ import React from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useAsync } from 'react-use';
 
@@ -78,6 +79,10 @@ const CatalogInfoAction = ({ data }: { data: AddRepositoryData }) => {
     values.repositories[data.id]?.catalogInfoYaml?.status ===
       RepositoryStatus.WAIT_PR_APPROVAL;
 
+  const canView =
+    values?.repositories?.[data.id]?.catalogInfoYaml?.status ===
+      RepositoryStatus.ADDED && values?.repositories?.[data.id]?.repoUrl;
+
   const removeQueryParams = () => {
     searchParams.delete('repository');
     searchParams.delete('defaultBranch');
@@ -110,20 +115,11 @@ const CatalogInfoAction = ({ data }: { data: AddRepositoryData }) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [repoUrl, defaultBranch, value?.status, values?.repositories, loading]);
 
-  return (
-    <Tooltip
-      title={
-        hasPermissionToEdit
-          ? 'Edit catalog-info.yaml pull request'
-          : 'View catalog-info.yaml file'
-      }
-    >
-      <span
-        data-testid={
-          hasPermissionToEdit ? 'edit-catalog-info' : 'view-catalog-info'
-        }
-      >
-        {hasPermissionToEdit ? (
+  const catalogIcon = () => {
+    if (hasPermissionToEdit) {
+      return {
+        tooltip: 'Edit catalog-info.yaml pull request',
+        icon: (
           <IconButton
             color="inherit"
             aria-label="Update"
@@ -133,25 +129,35 @@ const CatalogInfoAction = ({ data }: { data: AddRepositoryData }) => {
           >
             <EditIcon />
           </IconButton>
-        ) : (
+        ),
+        dataTestId: 'edit-catalog-info',
+      };
+    }
+    if (canView) {
+      return {
+        tooltip: 'View catalog-info.yaml file',
+        icon: (
           <IconButton
             target="_blank"
-            href={
-              values?.repositories?.[data.id]?.catalogInfoYaml?.prTemplate
-                ?.pullRequestUrl ||
-              values?.repositories?.[data.id]?.repoUrl ||
-              ''
-            }
+            href={canView}
             color="inherit"
             aria-label="View"
             size="large"
           >
             <OpenInNewIcon />
           </IconButton>
-        )}
-      </span>
+        ),
+        dataTestId: 'view-catalog-info',
+      };
+    }
+    return null;
+  };
+
+  return catalogIcon()?.tooltip ? (
+    <Tooltip title={catalogIcon()?.tooltip || ''}>
+      <span data-testid={catalogIcon()?.dataTestId}>{catalogIcon()?.icon}</span>
     </Tooltip>
-  );
+  ) : null;
 };
 
 export default CatalogInfoAction;