+++ workspaces/bulk-import/plugins/bulk-import/src/components/Repositories/RepositoriesList.test.tsx
@@ -12,7 +12,8 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */ import React from 'react';
+ */
+ import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 import { identityApiRef } from '@backstage/core-plugin-api';
@@ -64,7 +65,7 @@ const mockIdentityApi = {
 };
 
 const mockAsyncData = {
-  loaded: true,
+  loading: false,
   data: {
     addedRepositories: mockGetImportJobs.imports,
     totalJobs: mockGetImportJobs.imports.length,