+++ workspaces/bulk-import/plugins/bulk-import/src/utils/repository-utils.test.tsx
@@ -12,7 +12,8 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */ import {
+ */
+ import {
   mockGetRepositories,
   mockSelectedRepositories,
 } from '../mocks/mockData';
@@ -22,7 +23,6 @@ import {
   componentNameRegex,
   evaluatePRTemplate,
   getJobErrors,
-  getNewOrgsData,
   getYamlKeyValuePairs,
   updateWithNewSelectedRepositories,
   urlHelper,
@@ -78,33 +78,6 @@ describe('Repository utils', () => {
     expect(url).toBe('hjkh/hj');
     url = urlHelper('');
     expect(url).toBe('-');
-  });
-
-  it('should update organization data when repositories are selected', () => {
-    const newOrgsData = getNewOrgsData(
-      {
-        'org/dessert': {
-          id: '1234',
-          orgName: 'org/dessert',
-          organizationUrl: 'https://github.com/org/dessert',
-        },
-        'org/food': {
-          id: '1235',
-          orgName: 'org/food',
-          organizationUrl: 'https://github.com/org/food',
-        },
-        'org/pet-store-boston': {
-          id: '1236',
-          orgName: 'org/pet-store-boston',
-          organizationUrl: 'https://github.com/org/pet-store-boston',
-        },
-      },
-      mockGetRepositories.repositories[1],
-    );
-    expect(
-      Object.values(newOrgsData).find(o => o.orgName === 'org/dessert')
-        ?.selectedRepositories,
-    ).toEqual({ 'org/dessert/donut': mockGetRepositories.repositories[1] });
   });
 
   it('should parse key-value pairs correctly with semicolons', () => {