+++ workspaces/bulk-import/plugins/bulk-import/tests/bulkImport.spec.ts
@@ -12,7 +12,8 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */ import { expect, Page, test } from '@playwright/test';
+ */
+ import { expect, Page, test } from '@playwright/test';
 
 import { Common } from './bulkImportHelper';
 
@@ -117,6 +118,7 @@ test.describe('Bulk import plugin', () => {
       timeout: 20000,
     });
     await page.locator('button[aria-label="Next page"]').click();
+    await page.waitForTimeout(2000);
     await page.click('input[aria-label="select all repositories"]');
     await expect(
       page.getByRole('heading', { name: 'Selected repositories (9)' }),
@@ -124,6 +126,7 @@ test.describe('Bulk import plugin', () => {
       timeout: 20000,
     });
     await page.locator(`button`).filter({ hasText: 'Organization' }).click();
+    await page.waitForTimeout(2000);
     await expect(
       page.getByRole('heading', { name: 'Selected repositories (9)' }),
     ).toBeVisible({
@@ -139,6 +142,7 @@ test.describe('Bulk import plugin', () => {
 
   test('Select Repositories side panel is shown', async () => {
     await page.locator('button[type="button"][value="repository"]').click();
+    await page.waitForTimeout(2000);
     await expect(
       page.getByRole('heading', { name: 'Selected repositories (9)' }),
     ).toBeVisible({